requirements.txt
discord.py==1.7.3
discodo==3.0.5